Combine generator modules and tweak docs
tree.py appended to trees.py
inverse_line.py appended to line.py
