+++ doc/release/release_dev.rst
@@ -28,25 +28,28 @@ API Changes
 * [`#2498 <https://github.com/networkx/networkx/pull/2498>`_]
   Starting in NetworkX 2.1 the parameter ``alpha`` is deprecated and replaced
   with the customizable ``p_dist`` function parameter, which defaults to r^-2
-  if ``p_dist`` is not supplied. To reproduce networks of earlier NetworkX 
+  if ``p_dist`` is not supplied. To reproduce networks of earlier NetworkX
   versions, a custom function needs to be defined and passed as the ``p_dist``
   parameter. For example, if the parameter ``alpha`` = 2 was used in NetworkX 2.0,
   the custom function def custom_dist(r): r**-2 can be passed in versions >=2.1
-  as the parameter p_dist = custom_dist to produce an equivalent network. Note the
-  change in sign from +2 to -2 in this parameter change. 
+  as the parameter p_dist = custom_dist to produce an equivalent network.
+  Note the change in sign from +2 to -2 in this parameter change.
 
 * [`#2554 <https://github.com/networkx/networkx/issues/2554>`_]
   New algorithms for finding k-edge-connected components and k-edge-connected
   subgraphs in directed and undirected graphs. Efficient implementations are
   provided for the special case of k=1 and k=2. The new functionality is
   provided by:
-   - :func:`k_edge_components()`
-   - :func:`k_edge_subgraphs()`
+
+     :func:`k_edge_components()`
+
+     :func:`k_edge_subgraphs()`
 
 * [`#2572 <https://github.com/networkx/networkx/issues/2572>`_]
   New algorithm finding for finding k-edge-augmentations in undirected graphs.
   Efficient implementations are provided for the special case of k=1 and k=2.
   New functionality is provided by:
+
    - :func:`k_edge_augmentation()`
