+++ networkx/generators/trees.py
@@ -1,18 +1,19 @@
-# trees.py - functions for generating trees
+# -*- encoding: utf-8 -*-
+#    Copyright (C) 2015-2017 by
+#        Jeffrey Finkelstein <jeffrey.finkelstein@gmail.com>
+#        NetworkX developers
+#    All rights reserved.
+#    BSD license.
 #
-# Copyright 2016 Jeffrey Finkelstein <jeffrey.finkelstein@gmail.com>.
-#
-# This file is part of NetworkX.
-#
-# NetworkX is distributed under a BSD license; see LICENSE.txt for more
-# information.
-"""Provides functions for generating trees."""
+# Authors:      Jeffrey Finkelstein <jeffrey.finkelstein@gmail.com>
+"""Functions for generating trees."""
+import random
 from collections import defaultdict
 
 import networkx as nx
 from networkx.utils import generate_unique_node
 
-__all__ = ['prefix_tree']
+__all__ = ['prefix_tree', 'random_tree']
 
 #: The nil node, the only leaf node in a prefix tree.
 #:
@@ -95,7 +96,6 @@ def prefix_tree(paths):
         ['ab', 'abs', 'ad']
 
     """
-
     def _helper(paths, root, B):
         """Recursively create a trie from the given list of paths.
 
@@ -148,3 +148,51 @@ def prefix_tree(paths):
     # Populate the tree.
     _helper(paths, root, T)
     return T, root
+
+
+# From the Wikipedia article on Prüfer sequences:
+#
+# > Generating uniformly distributed random Prüfer sequences and
+# > converting them into the corresponding trees is a straightforward
+# > method of generating uniformly distributed random labelled trees.
+#
+def random_tree(n, seed=None):
+    """Returns a uniformly random tree on `n` nodes.
+
+    Parameters
+    ----------
+    n : int
+        A positive integer representing the number of nodes in the tree.
+
+    seed : int
+        A seed for the random number generator.
+
+    Returns
+    -------
+    NetworkX graph
+        A tree, given as an undirected graph, whose nodes are numbers in
+        the set {0, …, *n* - 1}.
+
+    Raises
+    ------
+    NetworkXPointlessConcept
+        If `n` is zero (because the null graph is not a tree).
+
+    Notes
+    -----
+    The current implementation of this function generates a uniformly
+    random Prüfer sequence then converts that to a tree via the
+    :func:`~networkx.from_prufer_sequence` function. Since there is a
+    bijection between Prüfer sequences of length *n* - 2 and trees on
+    *n* nodes, the tree is chosen uniformly at random from the set of
+    all trees on *n* nodes.
+
+    """
+    if n == 0:
+        raise nx.NetworkXPointlessConcept('the null graph is not a tree')
+    # Cannot create a Prüfer sequence unless `n` is at least two.
+    if n == 1:
+        return nx.empty_graph(1)
+    random.seed(seed)
+    sequence = [random.choice(range(n)) for i in range(n - 2)]
+    return nx.from_prufer_sequence(sequence)
